圖片

+++ app/Http/Controllers/FullCalenderController.php
@@ -14,7 +14,7 @@ class FullCalenderController extends Controller
         if(Auth::user()){
             if($request->ajax()){
 
-                $data = event::whereDate('start','>=',$request->start)->whereDate('end','<=',$request->end)->get();
+                $data = event::whereDate('start','>=',$request->start)->whereDate('end','<=',$request->end)->where('user_id','=',Auth::user()->id)->get();
                 return response()->json($data);
             }
             return view('index');
@@ -26,32 +26,36 @@ class FullCalenderController extends Controller
 
     public function action(Request $request){
 
-        if($request->ajax()){
-            if($request->type == 'add'){
-                $event = event::create([
-                    'title' => $request->title,
-                    'start' => $request->start,
-                    'end' => $request->end,
-                ]);
+        if(Auth::user()){
+            if($request->ajax()){
+                if($request->type == 'add'){
+                    $event = event::create([
+                        'title' => $request->title,
+                        'start' => $request->start,
+                        'end' => $request->end,
+                        'user_id' => Auth::user()->id,
+                    ]);
 
-                return response()->json($event);
-            }
+                    return response()->json($event);
+                }
 
-            if($request->type == 'update'){
-                $event = event::find($request->id)->update([
-                    'title' => $request->title,
-                    'start' => $request->start,
-                    'end' => $request->end,
-                ]);
-                return response()->json($event);
+                if($request->type == 'update'){
+                    $event = event::find($request->id)->update([
+                        'title' => $request->title,
+                        'start' => $request->start,
+                        'end' => $request->end,
+                    ]);
+                    return response()->json($event);
 
-            }
+                }
 
-            if($request->type == 'delete'){
-                $event = event::find($request->id)->delete();
+                if($request->type == 'delete'){
+                    $event = event::find($request->id)->delete();
 
-                return response()->json($event);
+                    return response()->json($event);
+                }
             }
         }
+
     }
 }
